Refactor pkg/networkfs
Solves #11591
