+++ daemon/networkdriver/bridge/driver.go
@@ -17,8 +17,8 @@ import (
 	"github.com/docker/docker/engine"
 	"github.com/docker/docker/nat"
 	"github.com/docker/docker/pkg/iptables"
-	"github.com/docker/docker/pkg/networkfs/resolvconf"
 	"github.com/docker/docker/pkg/parsers/kernel"
+	"github.com/docker/docker/pkg/resolvconf"
 	"github.com/docker/libcontainer/netlink"
 )
